Instead of disabling tests, merge test_internal package into test_util.
